extracted functionality to new project structure
