+++ src/database/database.module.ts
@@ -18,6 +18,7 @@ import { StatisticService } from './statistic.service';
 import { LanguageSchema } from './schemas/language.schema';
 import { CommitSchema } from './schemas/commit.schema';
 import { DeveloperFocus } from './statistics/developerFocus.service';
+import { FaultCorrection } from './statistics/faultCorrection.service';
 
 @Module({
   imports: [
@@ -39,7 +40,12 @@ import { DeveloperFocus } from './statistics/developerFocus.service';
       { name: 'Commit', schema: CommitSchema },
     ]),
   ],
-  providers: [DatabaseService, StatisticService, DeveloperFocus],
-  exports: [DatabaseService, StatisticService, DeveloperFocus],
+  providers: [
+    DatabaseService,
+    StatisticService,
+    DeveloperFocus,
+    FaultCorrection,
+  ],
+  exports: [DatabaseService, StatisticService, DeveloperFocus, FaultCorrection],
 })
 export class DatabaseModule {}
