+++ src/database/statistic.service.ts
@@ -737,192 +737,4 @@ export class StatisticService {
   wasCorrectedInTime(time: number, maximumDuration: number): boolean {
     return time > maximumDuration;
   }
-
-  /**
-   * The Fault Correction Rate describes the development team's capability to respond to bug reports.
-   * It is a quantitative indicator for which we consider all issues labeled bug (or some other equivalent tag) that existed at the time of a release.
-   * The Fault Correction Rate is the amount of closed bug issues divided by the total amount of bug issues.
-   *
-   * (release, issues) => {
-   * closed_bugs = issues[ label = bug, state = closed, closed_at <= release.created_at, closed_at >= release.previous().created_at ]
-   * open_bugs = issues[ label = bug, state = open, created_at <= release.created_at ]
-   * return |closed_bugs| / |closed_bugs| + |open_bugs|
-   *   }
-   * @param repoIdent
-   * @param userLimit
-   */
-  async faultCorrectionRate(
-    repoIdent: RepositoryNameDto,
-    labelNames?: string[],
-    userLimit?: number,
-  ) {
-    const limit = userLimit ? userLimit : 100;
-
-    const filter = {
-      repo: repoIdent.repo,
-      owner: repoIdent.owner,
-    };
-
-    const getIssuesWithEvents = {
-      from: 'issuewithevents',
-      localField: 'issuesWithEvents',
-      foreignField: '_id',
-      as: 'expandedIssuesWithEvents',
-    };
-
-    const getIssue = {
-      from: 'issues',
-      localField: 'expandedIssuesWithEvents.issue',
-      foreignField: '_id',
-      as: 'expandedIssue',
-    };
-
-    const getReleases = {
-      from: 'releases',
-      localField: 'releases',
-      foreignField: '_id',
-      as: 'expandedReleases',
-    };
-
-    const getLabel = {
-      from: 'labels',
-      localField: 'expandedIssue.label',
-      foreignField: '_id',
-      as: 'expandedIssue.expandedLabels',
-    };
-
-    //to obtain releases, sorted on the basis of created_at time
-    //TODO: limit amount of data returned, by using a projection
-    const releases: { _id: string; count: number }[] = await this.repoModel
-      .aggregate()
-      .match(filter)
-      .unwind('$releases')
-      .lookup(getReleases)
-      .unwind('$expandedReleases')
-      .sort({ 'expandedReleases.created_at': 1 })
-      //.limit(limit)
-      .exec();
-
-    //console.log(releases);
-
-    //to obtain closed issues, sorted on the basis of closed_at time
-    const issueQuery: any = this.repoModel
-      .aggregate()
-      .match(filter)
-      .project({ issuesWithEvents: 1 })
-      .unwind('$issuesWithEvents')
-      .lookup(getIssuesWithEvents)
-      .unwind('$expandedIssuesWithEvents')
-      .lookup(getIssue)
-      .unwind('$expandedIssue')
-      .lookup(getLabel)
-      .project({ expandedIssue: 1 });
-
-    if (labelNames) {
-      const match = labelNames.reduce(
-        (acc, curr) => {
-          acc.$and.push({
-            'expandedIssue.expandedLabels': { $elemMatch: { name: curr } },
-          });
-          return acc;
-        },
-        {
-          $and: [],
-        },
-      );
-      issueQuery.match(match);
-    }
-
-    const issues = await issueQuery.issueQuery
-      .sort({ 'expandedIssue.created_at': 1 })
-      .exec();
-
-    console.log(issues);
-    for (const issu of issues) {
-      for (const data of issu.expandedIssue.expandedLabels) {
-        console.log(data);
-      }
-    }
-    console.log(issues.length);
-    // //this loop matches the condition
-    // //closed_bugs = issues[ label = bug, state = closed, closed_at <= release.created_at, closed_at >= release.previous().created_at ]
-    // const closedBugs = this.getNoClosedBugs(closedIssues, releases);
-    // const openBugs = this.getNoOpenBugs(openIssues, releases);
-    // this.logger.debug(
-    //   `open bugs are: ${openBugs} and closed bugs are: ${closedBugs}`,
-    // );
-
-    // const fault_correction_rate =
-    //   Math.abs(closedBugs) / (Math.abs(closedBugs) + Math.abs(openBugs));
-
-    // this.logger.debug(`Fault correction rate is: ${fault_correction_rate}`);
-
-    // return fault_correction_rate;
-    return 0;
-  }
-
-  closedTicketsForRelease(
-    tickets: any[],
-    releases: any[],
-  ): { [release: string]: number } {
-    return {};
-  }
-  /**
-   *
-   * closed_bugs = issues[ label = bug, state = closed, closed_at <= release.created_at, closed_at >= release.previous().created_at ]
-   * @param closedIssues
-   * @param releases
-   * @returns
-   */
-  getNoClosedBugs(closedIssues: any[], releases: any[]): number {
-    let closedBugs = 0;
-    for (const closedIssue of closedIssues) {
-      for (let j = 1; j < releases.length; j++) {
-        if (
-          this.closedBeforeRelease(closedIssue, releases[j]) &&
-          this.closedAfterRelease(closedIssue, releases[j - 1])
-        ) {
-          closedBugs += 1;
-        }
-      }
-    }
-
-    return closedBugs;
-  }
-
-  getNoOpenBugs(openIssues: any[], releases: any[]): number {
-    //this loop matches below condition
-    //open_bugs = issues[ label = bug, state = open, created_at <= release.created_at ]
-    let openBugs = 0;
-    for (const openIssue of openIssues) {
-      for (let j = 1; j < releases.length; j++) {
-        if (this.openBeforeRelease(openIssue, releases[j])) {
-          openBugs += 1;
-          break;
-        }
-      }
-    }
-    return openBugs;
-  }
-
-  openBeforeRelease(issue, release) {
-    return (
-      issue['expandedIssue']['created_at'] <=
-      release['expandedReleases']['created_at']
-    );
-  }
-
-  closedBeforeRelease(issue, release) {
-    return (
-      issue['expandedIssue']['closed_at'] <=
-      release['expandedReleases']['created_at']
-    );
-  }
-
-  closedAfterRelease(issue, release) {
-    return (
-      issue['expandedIssue']['closed_at'] >=
-      release['expandedReleases']['created_at']
-    );
-  }
 }
